Cast the pickup-link id before it reaches the random-image query
PickupLink::get() returns me()->getPref("ipickup") verbatim, and setPref stores
whatever value it is handed -- it validates the preference name and not the
value, and api_set_preference takes it as check_arg type "string", which returns
the request value unmodified. min() then passed that string into an unquoted
numeric comparison, where escaping would not have helped anyway.

The sibling cookie entry is is_intger()-guarded, so omitting the cookie left the
stored string as the only input to min().

+++ offensive/pages/pic.php
@@ -54,7 +54,7 @@
 
 		$filter = me()->getPref("hide_nsfw") ? " AND nsfw = 0" : "";
 		$filter .= me()->getPref("hide_tmbo") ? " AND tmbo = 0" : "";
-		$sql = "SELECT id FROM offensive_uploads WHERE type='".$upload->type()."' AND status='normal' AND id < ".min($pickuplinks).$filter." ORDER BY RAND() LIMIT 1";
+		$sql = "SELECT id FROM offensive_uploads WHERE type='".$upload->type()."' AND status='normal' AND id < ".(int)min($pickuplinks).$filter." ORDER BY RAND() LIMIT 1";
 		$res = tmbo_query($sql);
 		$row = mysql_fetch_assoc( $res );
 		return($row['id']);
